mv grom logFormatter   int, int8, int16, int32, int64,  float32, float64 switch case orderby

+++ logs/logs.go
@@ -320,7 +320,7 @@ var logFormatter = func(values ...interface{}) (messages []interface{}) {
 						}
 					} else {
 						switch value.(type) {
-						case int, int16, int32, int64, int8, float32, float64:
+						case int, int8, int16, int32, int64, float32, float64:
 							formattedValues = append(formattedValues, fmt.Sprintf("%v", value))
 							break
 						default:
